@@ -15,7 +15,7 @@ dependencies:
   flutter_riverpod: ^2.5.1
   freezed_annotation: ^3.1.0
   json_annotation: ^4.9.0
-  supabase_flutter: ^2.6.0
+  supabase_flutter: ^2.9.1
   flutter_slidable: ^4.0.1
   flutter_svg: ^2.0.10+1
   intl: ^0.20.2
@@ -24,9 +24,9 @@ dependencies:
   uuid: ^4.4.0
   geolocator: any
   flutter_dotenv: any
+  build_runner: ^2.7.0
 
 dev_dependencies:
-  build_runner: ^2.4.11
   freezed: ^3.2.0
   json_serializable: ^6.9.0
   flutter_lints: ^6.0.0
