@@ -1,9 +1,11 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_stock/features/benefits/domain/user_benefit.dart';
 import 'package:flutter_stock/data/repositories/user_benefit_repository.dart';
+import 'package:flutter_stock/data/supabase/supabase_client_provider.dart';
+import 'package:flutter_stock/features/benefits/domain/user_benefit.dart';
 
 final userBenefitRepositoryProvider = Provider<UserBenefitRepository>((ref) {
-  return UserBenefitRepository();
+  final client = ref.watch(supabaseProvider);
+  return UserBenefitRepository(client);
 });
 
 final benefitListProvider = FutureProvider.autoDispose<List<UserBenefit>>((
