@@ -1,5 +1,5 @@
-import 'package:flutter_stock/features/benefits/domain/tag.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+
 part 'user_benefit.freezed.dart';
 part 'user_benefit.g.dart';
 
@@ -8,50 +8,28 @@ class UserBenefit with _$UserBenefit {
   const factory UserBenefit({
     required String id,
     required String userId,
-    required String companyCode,
-    required String companyName,
+    required String companyId, // ← これが無くて他所で落ちてた
+    String? companyName,
     required String benefitDetails,
     required DateTime expirationDate,
     @Default(false) bool isUsed,
-    @Default(<Tag>[]) List<Tag> tags,
-    // 任意のメモ。不要なら消してOK
     String? memo,
+    // tags は DB では別テーブル運用予定なので JSON は無視して OK
+    @Default(<Tag>[]) List<Tag> tags,
   }) = _UserBenefit;
 
   factory UserBenefit.fromJson(Map<String, dynamic> json) =>
       _$UserBenefitFromJson(json);
+}
 
-  @override
-  // TODO: implement benefitDetails
-  String get benefitDetails => throw UnimplementedError();
+DateTime _dtFromJson(dynamic v) =>
+    v is String ? DateTime.parse(v) : (v as DateTime);
 
-  @override
-  // TODO: implement companyCode
-  String get companyCode => throw UnimplementedError();
+dynamic _dtToJson(DateTime v) => v.toIso8601String();
 
-  @override
-  // TODO: implement companyName
-  String get companyName => throw UnimplementedError();
-
-  @override
-  // TODO: implement expirationDate
-  DateTime get expirationDate => throw UnimplementedError();
-
-  @override
-  // TODO: implement id
-  String get id => throw UnimplementedError();
-
-  @override
-  // TODO: implement isUsed
-  bool get isUsed => throw UnimplementedError();
-
-  @override
-  Map<String, dynamic> toJson() {
-    // TODO: implement toJson
-    throw UnimplementedError();
-  }
-
-  @override
-  // TODO: implement userId
-  String get userId => throw UnimplementedError();
+// 最小の Tag（警告/依存崩れ回避用）
+class Tag {
+  final String id;
+  final String name;
+  const Tag({required this.id, required this.name});
 }
