@@ -8,21 +8,23 @@ part of 'user_benefit.dart';
 
 _UserBenefit _$UserBenefitFromJson(Map<String, dynamic> json) => _UserBenefit(
   id: json['id'] as String,
-  userId: json['userId'] as String,
-  companyCode: json['companyCode'] as String,
-  companyName: json['companyName'] as String,
-  benefitDetails: json['benefitDetails'] as String,
-  expirationDate: DateTime.parse(json['expirationDate'] as String),
-  isUsed: json['isUsed'] as bool? ?? false,
+  userId: json['user_id'] as String,
+  companyId: json['company_id'] as String,
+  companyName: json['company_name'] as String?,
+  benefitDetails: json['benefit_details'] as String,
+  expirationDate: _dtFromJson(json['expiration_date']),
+  isUsed: json['is_used'] as bool? ?? false,
+  memo: json['memo'] as String?,
 );
 
 Map<String, dynamic> _$UserBenefitToJson(_UserBenefit instance) =>
     <String, dynamic>{
       'id': instance.id,
-      'userId': instance.userId,
-      'companyCode': instance.companyCode,
-      'companyName': instance.companyName,
-      'benefitDetails': instance.benefitDetails,
-      'expirationDate': instance.expirationDate.toIso8601String(),
-      'isUsed': instance.isUsed,
+      'user_id': instance.userId,
+      'company_id': instance.companyId,
+      'company_name': instance.companyName,
+      'benefit_details': instance.benefitDetails,
+      'expiration_date': _dtToJson(instance.expirationDate),
+      'is_used': instance.isUsed,
+      'memo': instance.memo,
     };
